+++ Engine/Source/Runtime/Datasmith/DatasmithCore/Private/DatasmithUtils.cpp
@@ -973,7 +973,7 @@ namespace DatasmithSceneUtilsImpl
 			{
 				if (ExpressionElement)
 				{
-					if (ExpressionElement->IsSubType((uint64)EDatasmithMaterialExpressionType::Texture))
+					if (ExpressionElement->IsSubType(EDatasmithMaterialExpressionType::Texture))
 					{
 						const IDatasmithMaterialExpressionTexture* TextureExpression = static_cast<IDatasmithMaterialExpressionTexture*>(ExpressionElement);
 						if (FCString::Strlen(TextureExpression->GetTexturePathName()) > 0)
@@ -985,7 +985,7 @@ namespace DatasmithSceneUtilsImpl
 							}
 						}
 					}
-					else if (ExpressionElement->IsSubType((uint64)EDatasmithMaterialExpressionType::Generic))
+					else if (ExpressionElement->IsSubType(EDatasmithMaterialExpressionType::Generic))
 					{
 						const IDatasmithMaterialExpressionGeneric* GenericExpression = static_cast<IDatasmithMaterialExpressionGeneric*>(ExpressionElement);
 
@@ -1000,7 +1000,7 @@ namespace DatasmithSceneUtilsImpl
 							}
 						}
 					}
-					else if (ExpressionElement->IsSubType((uint64)EDatasmithMaterialExpressionType::FunctionCall))
+					else if (ExpressionElement->IsSubType(EDatasmithMaterialExpressionType::FunctionCall))
 					{
 						const IDatasmithMaterialExpressionFunctionCall* FunctionExpression = static_cast<IDatasmithMaterialExpressionFunctionCall*>(ExpressionElement);
 						if (FCString::Strlen(FunctionExpression->GetFunctionPathName()) > 0)
@@ -1298,7 +1298,7 @@ namespace DatasmithSceneUtilsImpl
 				for (int32 AnimIndex = NumAnims - 1; AnimIndex >= 0; --AnimIndex)
 				{
 					TSharedPtr< IDatasmithBaseAnimationElement > Animation = LevelSequence->GetAnimation(AnimIndex);
-					if (Animation.IsValid() && Animation->IsA(EDatasmithElementType::Animation) && Animation->IsSubType((uint64)EDatasmithElementAnimationSubType::TransformAnimation))
+					if (Animation.IsValid() && Animation->IsA(EDatasmithElementType::Animation) && Animation->IsSubType(EDatasmithElementAnimationSubType::TransformAnimation))
 					{
 						const TSharedRef< IDatasmithTransformAnimationElement > TransformAnimation = StaticCastSharedRef< IDatasmithTransformAnimationElement >(Animation.ToSharedRef());
 
@@ -1340,12 +1340,12 @@ namespace DatasmithSceneUtilsImpl
 						continue;
 					}
 
-					if (AnimationElement->IsSubType((uint64)EDatasmithElementAnimationSubType::TransformAnimation))
+					if (AnimationElement->IsSubType(EDatasmithElementAnimationSubType::TransformAnimation))
 					{
 						const IDatasmithTransformAnimationElement* TransformAnimation = static_cast<IDatasmithTransformAnimationElement*>(AnimationElement.Get());
 						bValidSequence = ActorsInScene.Contains(TransformAnimation->GetName());
 					}
-					else if (AnimationElement->IsSubType((uint64)EDatasmithElementAnimationSubType::VisibilityAnimation))
+					else if (AnimationElement->IsSubType(EDatasmithElementAnimationSubType::VisibilityAnimation))
 					{
 						const IDatasmithVisibilityAnimationElement* VisibilityAnimation = static_cast<IDatasmithVisibilityAnimationElement*>(AnimationElement.Get());
 						bValidSequence = ActorsInScene.Contains(VisibilityAnimation->GetName());
@@ -1377,17 +1377,17 @@ namespace DatasmithSceneUtilsImpl
 						continue;
 					}
 
-					if (AnimationElement->IsSubType((uint64)EDatasmithElementAnimationSubType::TransformAnimation))
+					if (AnimationElement->IsSubType(EDatasmithElementAnimationSubType::TransformAnimation))
 					{
 						const IDatasmithTransformAnimationElement* TransformAnimation = static_cast<IDatasmithTransformAnimationElement*>(AnimationElement.Get());
 						bValidSequence = ActorsInScene.Contains(TransformAnimation->GetName());
 					}
-					else if (AnimationElement->IsSubType((uint64)EDatasmithElementAnimationSubType::VisibilityAnimation))
+					else if (AnimationElement->IsSubType(EDatasmithElementAnimationSubType::VisibilityAnimation))
 					{
 						const IDatasmithVisibilityAnimationElement* VisibilityAnimation = static_cast<IDatasmithVisibilityAnimationElement*>(AnimationElement.Get());
 						bValidSequence = ActorsInScene.Contains(VisibilityAnimation->GetName());
 					}
-					else if (AnimationElement->IsSubType((uint64)EDatasmithElementAnimationSubType::SubsequenceAnimation))
+					else if (AnimationElement->IsSubType(EDatasmithElementAnimationSubType::SubsequenceAnimation))
 					{
 						TSharedRef<IDatasmithSubsequenceAnimationElement> SubsequenceAnimation = StaticCastSharedRef<IDatasmithSubsequenceAnimationElement>(AnimationElement.ToSharedRef());
 						bValidSequence = ValidSequences.Contains(SubsequenceAnimation->GetSubsequence().Pin());
